Corrected a bug in the VideoMedia create method which was caused by changing relationships in the database.  Also tweaked the dashboard.
I added a step to the VideoMedia create method a while back that caused
it to create a Version record as well - an "Original Version."  This
step was broken by the introduction of additional non-nullable fields to
the data table.  I made the necessary modifications to fix the issue and
retain the functionality.