Added forward and backward buttons to the _Layout_Front partial view, allowing forward and backward navigation through Ajax request.
It seemed annoying that you couldn't just go back from one ajax request
to another, like you can with a less responsive design.  I added this
functionality to give the front-end user the benefits of going back and
forth like with more traditional browsing.  It's kind of like
breadcrumbs, but not.